refactor: merge forecast-preview-with-transparency into forecast-preview
Add optional transparency props to forecast-preview.tsx, default false.
Delete forecast-preview-with-transparency.tsx.

Phase 2C of codebase redundancy cleanup.
